requirements.txt
langchain
langchain-community
langchain-core
langchain-experimental
langchain-openai
langsmith
langchainhub
tiktoken
python-dotenv
chromadb
requests
termcolor
langchain-chroma
